api/main/requirement.txt
flask
flasgger
Keras==2.4.3
Pillow
